requirements.txt
kafka-python==2.3.0
boto3==1.42.80
botocore==1.42.80
requests==2.32.3
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.0
fredapi==0.5.0
alpha_vantage==2.3.1
snowflake-connector-python==3.7.0
dbt-core==1.11.7
dbt-snowflake==1.11.3
great_expectations==0.18.0
streamlit==1.32.0
plotly==5.19.0
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0
apache-airflow==2.8.1
confluent-kafka==2.3.0
pyarrow==15.0.0
pytest==8.0.0
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.19.0
kafka-python-ng>=2.2.0
